requirements.txt
openai>=1.0.0
google-generativeai>=0.8.0
pydub>=0.25.1
edge-tts>=6.1.0
deep-translator>=1.11.0
soundfile>=0.12.0
numpy>=1.24.0,<2.0.0
aiofiles>=23.0.0
python-dotenv>=1.0.0 